backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
sqlalchemy==2.0.46
pydantic==2.12.5
pyjwt==2.11.0
httpx==0.28.1
